BIVOT_PSEUDO_CODE V2
Added more detail to the calibration section. It would be good to up load if I had pin info and how to read from adc pins. No update on battery management code look like it might be pushed back to V2 of the prototype. Add sensor functionality which just checks if the sensors are in marginal range. Read state of valve was add as part of main code and should later be moved into a function for organization.  All that is left to do is the store and forward function, the state transmission, and the set sleep mode.